+++ src/code42cli/cmds/detectionlists/enums.py
@@ -1,6 +1,6 @@
 class DetectionLists(object):
     DEPARTING_EMPLOYEE = u"departing-employee"
-    HIGH_RISK = u"high-risk"
+    HIGH_RISK_EMPLOYEE = u"high-risk-employee"
 
 
 class BulkCommandType(object):
@@ -8,3 +8,10 @@ class BulkCommandType(object):
 
     def __iter__(self):
         return iter([self.ADD])
+
+
+class DetectionListUserKeys(object):
+    CLOUD_ALIAS = u"cloud_alias"
+    USERNAME = u"username"
+    NOTES = u"notes"
+    RISK_TAG = u"risk_tag"